fix(experiments): check smoke convergence elementwise, widen float fallback
Review feedback on #61.

- harness_smoke.jl: two different vectors can share a norm, so the norm ratio
  alone could not prove convergence. Monotonicity, the discrete upper bound,
  and convergence are now checked component by component on the readout
  vectors, and metrics.csv gains max_abs_deviation / max_rel_deviation columns
  (max elementwise deviation is 0.0016 of the peak component at tau = 200).
- ExperimentUtils: the _format_float fallback now widens to Float64 instead of
  narrowing to Float32. Float32 and Float64 already dispatch to the concrete
  method and keep full precision; this only affects other float types such as
  Float16 and BigFloat.

+++ experiments/src/ExperimentUtils.jl
@@ -187,7 +187,9 @@ end
 
 write_config(slug::AbstractString, cfg::NamedTuple) = write_config(slug, Dict(pairs(cfg)))
 
-_format_float(x::AbstractFloat) = _format_float(Float32(x))
+# Float32 and Float64 hit the concrete method below and keep full precision;
+# any other float type (Float16, BigFloat, ...) is widened, never narrowed.
+_format_float(x::AbstractFloat) = _format_float(Float64(x))
 function _format_float(x::Union{Float32,Float64})
     isnan(x) && return "NaN"
     isinf(x) && return x > 0 ? "Inf" : "-Inf"
